move function to util per pr suggestion
