+++ tests/modules/encounters/resources/test_modify_encounter.py
@@ -1,30 +1,8 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=missing-docstring
-import json
 from tests import utils
 from tests.modules.encounters.resources import utils as enc_utils
 from tests.extensions.edm import utils as edm_utils
-
-
-PATH = '/api/v1/encounters/'
-
-
-def patch_encounter(
-    flask_app_client, encounter_guid, user, data, expected_status_code=200
-):
-    with flask_app_client.login(user, auth_scopes=('encounters:write',)):
-        response = flask_app_client.patch(
-            '%s%s' % (PATH, encounter_guid),
-            content_type='application/json',
-            data=json.dumps(data),
-        )
-    if expected_status_code == 200:
-        utils.validate_dict_response(response, 200, {'id', 'version'})
-    else:
-        utils.validate_dict_response(
-            response, expected_status_code, {'status', 'message'}
-        )
-    return response
 
 
 def test_modify_encounter(db, flask_app_client, researcher_1, researcher_2, admin_user):
@@ -41,7 +19,7 @@ def test_modify_encounter(db, flask_app_client, researcher_1, researcher_2, admi
         utils.patch_replace_op('owner', str(researcher_1.guid)),
         utils.patch_replace_op('locationId', 'FAIL'),
     ]
-    res = patch_encounter(
+    res = enc_utils.patch_encounter(
         flask_app_client,
         new_encounter_1.guid,
         researcher_1,
@@ -55,7 +33,7 @@ def test_modify_encounter(db, flask_app_client, researcher_1, researcher_2, admi
         utils.patch_replace_op('owner', str(researcher_2.guid)),
     ]
 
-    patch_encounter(
+    enc_utils.patch_encounter(
         flask_app_client,
         new_encounter_1.guid,
         researcher_2,
@@ -69,7 +47,7 @@ def test_modify_encounter(db, flask_app_client, researcher_1, researcher_2, admi
         utils.patch_test_op(researcher_1.password_secret),
         utils.patch_replace_op('owner', str(researcher_2.guid)),
     ]
-    patch_encounter(
+    enc_utils.patch_encounter(
         flask_app_client, new_encounter_1.guid, researcher_1, new_owner_as_res_1
     )
     assert new_encounter_1.owner == researcher_2
@@ -77,7 +55,7 @@ def test_modify_encounter(db, flask_app_client, researcher_1, researcher_2, admi
     # test changing locationId via patch
     new_val = 'LOCATION_TEST_VALUE'
     patch_data = [utils.patch_replace_op('locationId', new_val)]
-    res = patch_encounter(
+    res = enc_utils.patch_encounter(
         flask_app_client, new_encounter_1.guid, researcher_2, patch_data
     )
     assert res is not None
@@ -101,7 +79,7 @@ def test_modify_encounter(db, flask_app_client, researcher_1, researcher_2, admi
             'customFields', {'id': cfd_id, 'value': new_cfd_test_value}
         )
     ]
-    response = patch_encounter(
+    response = enc_utils.patch_encounter(
         flask_app_client, new_encounter_1.guid, researcher_2, patch_data
     )
