+++ tests/modules/encounters/resources/utils.py
@@ -4,6 +4,7 @@ Project resources utils
 -------------
 """
 from tests import utils as test_utils
+import json
 
 PATH = '/api/v1/encounters/'
 
@@ -20,6 +21,24 @@ def create_encounter(flask_app_client, user, expected_status_code=200):
     response = sighting_utils.create_sighting(flask_app_client, user, data_in)
     assert isinstance(response.json, dict)
     assert response.status_code == expected_status_code
+    return response
+
+
+def patch_encounter(
+    flask_app_client, encounter_guid, user, data, expected_status_code=200
+):
+    with flask_app_client.login(user, auth_scopes=('encounters:write',)):
+        response = flask_app_client.patch(
+            '%s%s' % (PATH, encounter_guid),
+            content_type='application/json',
+            data=json.dumps(data),
+        )
+    if expected_status_code == 200:
+        test_utils.validate_dict_response(response, 200, {'id', 'version'})
+    else:
+        test_utils.validate_dict_response(
+            response, expected_status_code, {'status', 'message'}
+        )
     return response
